+++ CHANGELOG.md
@@ -6,11 +6,13 @@
 
 - Get the exports of the chunks in next.js
 - Added command to get the exports of a chunk in next.js - `list exportnames`
+- Added `--mapped-json` flag to endpoints module
 
 ### Changed
 
 - Type for `exports` is now `string[]` rather than `string` in `mapped.json`
 - Remove `md` output format for endpoints module
+- Removed `--subsequent-requests-dir` option from endpoints module (now use `-d` only)
 
 ### Fixed